+++ py2app_tests/basic_plugin/main.py
@@ -3,5 +3,5 @@ import Foundation
 
 class BasicPlugin (Foundation.NSObject):
     def performCommand_(self, cmd):
-        print "+ %s"%(cmd,)
+        print ("+ %s"%(cmd,))
         sys.stdout.flush()